Factor out Read.Block.Txs from Read.Block
