+++ lib/read/lib/Cardano/Wallet/Read/Block.hs
@@ -1,9 +1,6 @@
-{-# LANGUAGE ExistentialQuantification #-}
-{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
-{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 
 -- |
@@ -16,7 +13,6 @@ module Cardano.Wallet.Read.Block
     ( ConsensusBlock
     , Block (..)
     , fromConsensusBlock
-    , getTxs
     , toConsensusBlock
     ) where
 
@@ -34,31 +30,18 @@ import Cardano.Api
 import Cardano.Ledger.Api
     ( StandardCrypto
     )
-import Cardano.Ledger.Binary
-    ( EncCBOR
-    )
 import Cardano.Wallet.Read.Eras
-    ( (:.:) (..)
-    , EraFun (..)
+    ( EraFun (..)
     , EraValue
     , K (..)
     , allegra
     , alonzo
-    , applyEraFun
     , babbage
     , byron
     , conway
     , inject
     , mary
-    , sequenceEraValue
     , shelley
-    )
-import Cardano.Wallet.Read.Tx
-    ( Tx (..)
-    , TxT
-    )
-import Data.Foldable
-    ( toList
     )
 import Ouroboros.Consensus.Protocol.Praos
     ( Praos
@@ -66,23 +49,11 @@ import Ouroboros.Consensus.Protocol.Praos
 import Ouroboros.Consensus.Protocol.TPraos
     ( TPraos
     )
-import Ouroboros.Consensus.Shelley.Protocol.Abstract
-    ( ShelleyProtocolHeader
-    )
 
-import qualified Cardano.Chain.Block as Byron
-import qualified Cardano.Chain.UTxO as Byron
-import qualified Cardano.Ledger.Api as Ledger
-import qualified Cardano.Ledger.Era as Shelley
-import qualified Cardano.Ledger.Shelley.API as Shelley
-import qualified Ouroboros.Consensus.Byron.Ledger as Byron
 import qualified Ouroboros.Consensus.Byron.Ledger as O
 import qualified Ouroboros.Consensus.Cardano.Block as O
 import qualified Ouroboros.Consensus.Shelley.Ledger as O
 
-{-------------------------------------------------------------------------------
-    Block type
--------------------------------------------------------------------------------}
 -- | Type synonym for 'CardanoBlock' with cryptography as used on mainnet.
 type ConsensusBlock = O.CardanoBlock O.StandardCrypto
 
@@ -107,35 +78,6 @@ newtype Block era = Block {unBlock :: BlockT era}
 deriving instance Show (BlockT era) => Show (Block era)
 deriving instance Eq (BlockT era) => Eq (Block era)
 
--- | Get sequence of transactions in the block.
-txsFromBlockE :: EraFun Block ([] :.: Tx)
-txsFromBlockE =
-    EraFun
-        { byronFun = getTxs' getTxsFromBlockByron
-        , shelleyFun = getTxs' getTxsFromBlockShelleyAndOn
-        , maryFun = getTxs' getTxsFromBlockShelleyAndOn
-        , allegraFun = getTxs' getTxsFromBlockShelleyAndOn
-        , alonzoFun = getTxs' getTxsFromBlockShelleyAndOn
-        , babbageFun = getTxs' getTxsFromBlockShelleyAndOn
-        , conwayFun = getTxs' getTxsFromBlockShelleyAndOn
-        }
-  where
-    getTxs' f (Block block) = Comp $ Tx <$> f block
-
-getTxsFromBlockByron :: O.ByronBlock -> [TxT ByronEra]
-getTxsFromBlockByron block =
-    case Byron.byronBlockRaw block of
-        Byron.ABOBBlock b ->
-            map (() <$) . Byron.unTxPayload . Byron.blockTxPayload $ b
-        Byron.ABOBBoundary _ -> []
-
-getTxsFromBlockShelleyAndOn
-    :: (Shelley.EraSegWits era, EncCBOR (ShelleyProtocolHeader proto))
-    => O.ShelleyBlock proto era
-    -> [Ledger.Tx era]
-getTxsFromBlockShelleyAndOn (O.ShelleyBlock (Shelley.Block _ txs) _) =
-    toList (Shelley.fromTxSeq txs)
-
 -- | Convert block as received from cardano-node
 -- via Haskell library of mini-protocol.
 fromConsensusBlock :: ConsensusBlock -> EraValue Block
@@ -148,16 +90,14 @@ fromConsensusBlock = \case
     O.BlockBabbage block -> inject babbage $ Block block
     O.BlockConway block -> inject conway $ Block block
 
-getTxs :: O.CardanoBlock StandardCrypto -> [EraValue Tx]
-getTxs = sequenceEraValue . applyEraFun txsFromBlockE . fromConsensusBlock
-
 toConsensusBlock :: EraFun Block (K ConsensusBlock)
-toConsensusBlock = EraFun
-    { byronFun = K . O.BlockByron . unBlock
-    , shelleyFun = K . O.BlockShelley . unBlock
-    , allegraFun = K . O.BlockAllegra . unBlock
-    , maryFun = K . O.BlockMary . unBlock
-    , alonzoFun = K . O.BlockAlonzo . unBlock
-    , babbageFun = K . O.BlockBabbage . unBlock
-    , conwayFun = K . O.BlockConway . unBlock
-    }
+toConsensusBlock =
+    EraFun
+        { byronFun = K . O.BlockByron . unBlock
+        , shelleyFun = K . O.BlockShelley . unBlock
+        , allegraFun = K . O.BlockAllegra . unBlock
+        , maryFun = K . O.BlockMary . unBlock
+        , alonzoFun = K . O.BlockAlonzo . unBlock
+        , babbageFun = K . O.BlockBabbage . unBlock
+        , conwayFun = K . O.BlockConway . unBlock
+        }
