+++ lib/read/lib/Cardano/Wallet/Read/Block/Txs.hs
@@ -1,0 +1,95 @@
+{-# LANGUAGE ExistentialQuantification #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
+
+module Cardano.Wallet.Read.Block.Txs
+    ( getTxs
+    , getTxsForEra
+    , txsFromBlock
+    ) where
+
+import Prelude
+
+import Cardano.Api
+    ( ByronEra
+    )
+import Cardano.Ledger.Api
+    ( StandardCrypto
+    )
+import Cardano.Ledger.Binary
+    ( EncCBOR
+    )
+import Cardano.Wallet.Read.Block
+    ( Block (..)
+    , fromConsensusBlock
+    )
+import Cardano.Wallet.Read.Eras
+    ( (:.:) (..)
+    , EraFun (..)
+    , EraValue
+    , MkEraValue
+    , applyEraFun
+    , project
+    )
+import Cardano.Wallet.Read.Tx
+    ( Tx (..)
+    , TxT
+    )
+import Data.Foldable
+    ( toList
+    )
+import Generics.SOP
+    ( unComp
+    )
+import Ouroboros.Consensus.Shelley.Protocol.Abstract
+    ( ShelleyProtocolHeader
+    )
+
+import qualified Cardano.Chain.Block as Byron
+import qualified Cardano.Chain.UTxO as Byron
+import qualified Cardano.Ledger.Api as Ledger
+import qualified Cardano.Ledger.Era as Shelley
+import qualified Cardano.Ledger.Shelley.API as Shelley
+import qualified Ouroboros.Consensus.Byron.Ledger as Byron
+import qualified Ouroboros.Consensus.Byron.Ledger as O
+import qualified Ouroboros.Consensus.Cardano.Block as O
+import qualified Ouroboros.Consensus.Shelley.Ledger as O
+
+-- | Get sequence of transactions in the block.
+txsFromBlock :: EraFun Block ([] :.: Tx)
+txsFromBlock =
+    EraFun
+        { byronFun = getTxs' getTxsFromBlockByron
+        , shelleyFun = getTxs' getTxsFromBlockShelleyAndOn
+        , maryFun = getTxs' getTxsFromBlockShelleyAndOn
+        , allegraFun = getTxs' getTxsFromBlockShelleyAndOn
+        , alonzoFun = getTxs' getTxsFromBlockShelleyAndOn
+        , babbageFun = getTxs' getTxsFromBlockShelleyAndOn
+        , conwayFun = getTxs' getTxsFromBlockShelleyAndOn
+        }
+  where
+    getTxs' f (Block block) = Comp $ Tx <$> f block
+
+getTxsFromBlockByron :: O.ByronBlock -> [TxT ByronEra]
+getTxsFromBlockByron block =
+    case Byron.byronBlockRaw block of
+        Byron.ABOBBlock b ->
+            map (() <$) . Byron.unTxPayload . Byron.blockTxPayload $ b
+        Byron.ABOBBoundary _ -> []
+
+getTxsFromBlockShelleyAndOn
+    :: (Shelley.EraSegWits era, EncCBOR (ShelleyProtocolHeader proto))
+    => O.ShelleyBlock proto era
+    -> [Ledger.Tx era]
+getTxsFromBlockShelleyAndOn (O.ShelleyBlock (Shelley.Block _ txs) _) =
+    toList (Shelley.fromTxSeq txs)
+
+getTxs :: O.CardanoBlock StandardCrypto -> EraValue ([] :.: Tx)
+getTxs = applyEraFun txsFromBlock . fromConsensusBlock
+
+getTxsForEra
+    :: MkEraValue ([] :.: Tx) p
+    -> O.CardanoBlock StandardCrypto
+    -> Maybe [Tx p]
+getTxsForEra ew = fmap unComp . project ew . getTxs
